requirements.txt
fastapi
uvicorn
python-multipart
kafka-python
redis
qdrant-client
falkordb
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai
langchain-ollama
langchain-text-splitters>=0.3.0
aiohttp
tiktoken
aiohttp
tenacity
python-dotenv
unstructured
python-magic
pypdf
langchain-google-genai
gradio
requests
flashrank